Replaced Defined routing with Handles routing
The Handles routing is basically the same as Defined but does not take
params into account when finding the requested path.

Basically, POST /c/a --> controller@action. Nothing more than that. You
are replacing the long URL with a short one. This is basically an
aliasing system with HTTP verbs